DEV-1193: add Skill.from_file() classmethod
Extract pure parsing/validation helpers into _parsing.py to break the
circular-import chain, then expose Skill.from_file() as the canonical
single-skill loader. The private _load_skill_from_file() in directory.py
is removed; discover_skills() now delegates to Skill.from_file().

With validate=True a missing name field now raises SkillValidationError
instead of silently returning None, making the behaviour consistent.
